+++ src/lexer/model/token.ts
@@ -14,22 +14,51 @@ export type TokenDataBinding<Kinds extends string, Data> = {
 
 export type GeneralTokenDataBinding = TokenDataBinding<string, unknown>;
 
+export type Range = {
+  /**
+   * 0-based index.
+   */
+  start: number;
+  /**
+   * 0-based index. Exclusive.
+   */
+  end: number;
+};
+
 /**
  * The output of a lexer.
  */
-export type Token<DataBindings extends GeneralTokenDataBinding, ErrorType> = {
-  /**
-   * Token's text content.
-   */
-  content: string;
-  /**
-   * Index of the first char of this token in the whole input string.
-   */
-  start: number;
+export class Token<DataBindings extends GeneralTokenDataBinding, ErrorType> {
+  kind: DataBindings["kind"];
+  data: DataBindings["data"];
+  buffer: string;
+  range: Range;
   /**
    * `undefined` means no error.
    */
   error?: ErrorType;
-} & DataBindings;
+
+  private constructor() {}
+
+  static from<Kind extends string, Data, ErrorType>(
+    kind: Kind,
+    data: Data,
+    buffer: string,
+    range: Range,
+    error?: ErrorType,
+  ): Token<{ kind: Kind; data: Data }, ErrorType> {
+    const t = new Token<{ kind: Kind; data: Data }, ErrorType>();
+    t.kind = kind;
+    t.data = data;
+    t.buffer = buffer;
+    t.range = range;
+    t.error = error;
+    return t;
+  }
+
+  get content(): string {
+    return this.buffer.slice(this.range.start, this.range.end);
+  }
+}
 
 export type GeneralToken = Token<GeneralTokenDataBinding, unknown>;